+++ lib/ast-transform.js
